JS
This code is JavaScript that is used to generate a random password based on certain user-specified parameters. It defines a number of functions that generate a random lowercase letter, uppercase letter, number, and symbol. These functions are later used to generate the password. It also defines an object that maps the names of these functions to their corresponding function. An event listener is added to the generate button element, which when clicked, calls the generatePassword function, passing in the values of the user-specified settings and the desired length of the password. The generatePassword function uses the passed in values to determine what types of characters to include in the generated password, and uses the previously defined functions to generate a random password of the specified length. It also has a click event listener on the copy button element, which when clicked, copies the generated password to the clipboard.